[value entity] renamed tck directory.
