+++ tck/value/model.go
